Closed #22.

+++ examples/03-Analog/scope-analog/render.cpp
@@ -28,7 +28,7 @@ The Bela software is distributed under the GNU Lesser General Public License
 
 Scope scope;
 
-float gAudioFramesPerAnalogFrame;
+int gAudioFramesPerAnalogFrame;
 float gInverseSampleRate;
 float gPhase;
 
@@ -37,6 +37,9 @@ float gFrequency;
 
 float gIn1;
 float gIn2;
+
+// For this example you need to set the Analog Sample Rate to 
+// 44.1 KHz which you can do in the settings tab.
 
 bool setup(BelaContext *context, void *userData)
 {
